Rename files and minor cleanup
Renames log to progress_bar, cluster to filepath, and moves leftover
functionality from db file to util file

Also, rename BackupDirMap to SegDirMap
