+++ backup/queries_globals.go
@@ -308,3 +308,11 @@ AND spcname != 'pg_global';`
 	gplog.FatalOnError(err)
 	return results
 }
+
+func GetDBSize(connection *dbconn.DBConn) string {
+	size := struct{ DBSize string }{}
+	sizeQuery := fmt.Sprintf("SELECT pg_size_pretty(sodddatsize) as dbsize FROM gp_toolkit.gp_size_of_database WHERE sodddatname=E'%s'", dbconn.EscapeConnectionParam(connection.DBName))
+	err := connection.Get(&size, sizeQuery)
+	gplog.FatalOnError(err)
+	return size.DBSize
+}
